pyproject.toml
[tool.poetry]
name = "extension-backend-api"
version = "0.1.0"
description = ""
authors = ["jeahoon.jang <jaehoon.jang@fitpet.co.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
sqlalchemy = "^2.0.40"
marshmallow = "^3.26.1"
flask-sqlalchemy = "^3.1.1"
flask-marshmallow = "^1.3.0"
marshmallow-sqlalchemy = "^1.4.1"
pymysql = "^1.1.1"
flask-cors = "^5.0.1"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
